+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 - Changed style of the inputs when editing _lithological_descriptions_ and _facies_descriptions_ to _outlined_.
+- Borehole duplicates are only identified within the same workgroup when importing boreholes
 
 ## v2.0.298 - 2023-04-18